requirements.txt
django
psycopg2-binary
gunicorn
django-heroku
whitenoise
dj-database-url
django-cors-headers
djangorestframework
cryptography
django-rest-knox
rest-social-auth
drf-nested-routers
requests
python-twitter
feedparser
python-dateutil
coverage
